Kernel: Xiaomi kernel changes for Red Note5A.
The Patch based on QualComm release version:LA.UM.5.6.r1-04600-89xx.0.
The kernel config file used is ugg-msm8937_defconfig/ugglite-_msm8937_defconfig

+++ include/linux/mmc/card.h
@@ -106,7 +106,7 @@ struct mmc_ext_csd {
 	u8			raw_ext_csd_structure;	/* 194 */
 	u8			raw_card_type;		/* 196 */
 	u8			raw_drive_strength;	/* 197 */
-	u8			out_of_int_time;	/* 198 */
+	u16			out_of_int_time;	/* 198 */
 	u8			raw_pwr_cl_52_195;	/* 200 */
 	u8			raw_pwr_cl_26_195;	/* 201 */
 	u8			raw_pwr_cl_52_360;	/* 202 */
